# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EDCS API",
        "description": "The Elite Dangerous Companion Suite API provides galaxy data from Elite Dangerous to various consumers, including the EDCS frontend at https://edcs.app. Data is sourced from EDSM (Elite Dangerous Star Map), EDDN (Elite Dangerous Data Network), and the Frontier Companion API (CAPI).",
        "contact": {
            "name": "EDCS",
            "url": "https://edcs.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.edcs.app/api",
            "description": "EDCS API Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user account",
                "description": "Register users",
                "operationId": "6703d7156c265a1e7b47e500d2a16304",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chris Rowles"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "me@rowles.ch"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered — returns a Sanctum bearer token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthToken"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in with email and password",
                "description": "Login users",
                "operationId": "a5385ffca73f113abef00ea604d1ca58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "me@rowles.ch"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated — returns a Sanctum bearer token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthToken"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke all tokens for the authenticated user",
                "description": "Logout users",
                "operationId": "0f6715df0d22b69c7e35bd32b4c1fc92",
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user with commander details",
                "description": "Fetch authenticated user",
                "operationId": "37252cdc2ae05f1e2e4fd541e987489d",
                "responses": {
                    "200": {
                        "description": "Authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/auth/frontier/callback": {
            "get": {
                "tags": [
                    "Frontier Auth"
                ],
                "summary": "Handle Frontier SSO OAuth callback",
                "description": "Exchanges the authorization code for a Frontier access token, creates/confirms the user and commander record, then redirects to the frontend with a cmdr_token HttpOnly cookie.",
                "operationId": "5d38b253137ef7813c0e7b1e25635ae6",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Authorization grant code from Frontier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code_verifier",
                        "in": "query",
                        "description": "PKCE code verifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to frontend with cmdr_token cookie set"
                    },
                    "400": {
                        "description": "Invalid authorization code or verifier"
                    }
                }
            }
        },
        "/auth/frontier/me": {
            "post": {
                "tags": [
                    "Frontier Auth"
                ],
                "summary": "Get the Frontier-authenticated user via cookie",
                "description": "Returns the authenticated user and their commander data using the cmdr_token HttpOnly cookie set during the SSO callback.",
                "operationId": "508608da4dc2df7e42f001057de28c84",
                "responses": {
                    "200": {
                        "description": "Authenticated user with commander data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated — missing or invalid cmdr_token cookie"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthToken": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Chris Rowles"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|abc123..."
                    },
                    "expiry": {
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "Commander": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "CMDR Rowles"
                    },
                    "api": {
                        "properties": {
                            "edsm": {
                                "type": "string",
                                "nullable": true
                            },
                            "inara": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Chris Rowles"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "me@rowles.ch"
                    },
                    "commander": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Commander"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum bearer token. Obtain via POST /auth/login or the Frontier SSO flow.",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}